refactor tcp connection setup into webserver class
